Move ParserOptions from common.hpp to new header parse_options.hpp
