+++ src/c4/yml/common.hpp
@@ -4,9 +4,16 @@
 /** @file common.hpp Common utilities and infrastructure used by ryml. */
 
 #include <cstddef>
+
+#ifndef _C4_SUBSTR_HPP_
 #include <c4/substr.hpp>
+#endif
+#ifndef _C4_CHARCONV_HPP_
 #include <c4/charconv.hpp>
+#endif
+#ifndef _C4_YML_EXPORT_HPP_
 #include <c4/yml/export.hpp>
+#endif
 
 
 //-----------------------------------------------------------------------------
@@ -342,118 +349,6 @@ struct RYML_EXPORT ErrorDataVisit
     id_type node;     ///< node where the error was detected
     ErrorDataVisit() noexcept = default;
     ErrorDataVisit(Location const& cpploc_, Tree const *tree_ , id_type node_) noexcept : cpploc(cpploc_), tree(tree_), node(node_) {}
-};
-
-
-
-//-----------------------------------------------------------------------------
-//-----------------------------------------------------------------------------
-//-----------------------------------------------------------------------------
-
-/** Options to give to the parser to control its behavior. */
-struct RYML_EXPORT ParserOptions
-{
-private:
-
-    typedef enum : uint32_t { // NOLINT
-        DETECT_FLOW_ML = (1u << 0u),
-        RESOLVE_TAGS = (1u << 1u),
-        RESOLVE_TAGS_ALL = (1u << 2u),
-        SCALAR_FILTERING = (1u << 3u),
-        LOCATIONS = (1u << 4u),
-        DEFAULTS = SCALAR_FILTERING|DETECT_FLOW_ML,
-    } Flags_e;
-
-    uint32_t flags = DEFAULTS;
-
-    ParserOptions& set_flags_(bool enabled, Flags_e f)
-    {
-        if(enabled)
-            flags |= f;
-        else
-            flags &= ~f;
-        return *this;
-    }
-
-public:
-
-    ParserOptions() = default;
-
-public:
-
-    /** @name detection of @ref FLOW_ML container style */
-    /** @{ */
-
-    /** enable/disable detection of @ref FLOW_ML container style. When
-     * enabled, the parser will set @ref FLOW_ML as the style of flow
-     * containers which have the terminating bracket on a line
-     * different from that of the opening bracket. */
-    ParserOptions& detect_flow_ml(bool enabled) noexcept
-    {
-        return set_flags_(enabled, DETECT_FLOW_ML);
-    }
-    /** query status of detection of @ref FLOW_ML container style. */
-    C4_ALWAYS_INLINE bool detect_flow_ml() const noexcept { return (flags & DETECT_FLOW_ML); }
-
-    /** @} */
-
-public:
-
-    /** @name resolution of tags */
-    /** @{ */
-
-    /** enable/disable resolution of YAML tags during parsing. When
-     * enabled, tags are resolved according to existing tag
-     * directives. Disabled by default. See also @ref
-     * ParserOptions::resolve_tags_all(). */
-    ParserOptions& resolve_tags(bool enabled) noexcept
-    {
-        return set_flags_(enabled, RESOLVE_TAGS);
-    }
-    /** query status of tag resolution setting. */
-    C4_ALWAYS_INLINE bool resolve_tags() const noexcept { return (flags & RESOLVE_TAGS); }
-
-    /** When resolve_tags() is enabled, resolve not just prefixed tags
-     * of the form <pre>!handle!tag</pre>, but also non-prefixed tags
-     * (<pre>!!tag</pre> and <pre>!tag!</pre>). Disabled by default. */
-    ParserOptions& resolve_tags_all(bool enabled) noexcept
-    {
-        return set_flags_(enabled, RESOLVE_TAGS_ALL);
-    }
-    /** query status of non-prefixed tag resolution setting. */
-    C4_ALWAYS_INLINE bool resolve_tags_all() const noexcept { return (flags & RESOLVE_TAGS_ALL); }
-
-    /** @} */
-
-public:
-
-    /** @name source location tracking */
-    /** @{ */
-
-    /** enable/disable source location tracking */
-    ParserOptions& locations(bool enabled) noexcept
-    {
-        return set_flags_(enabled, LOCATIONS);
-    }
-    /** query source location tracking status */
-    C4_ALWAYS_INLINE bool locations() const noexcept { return (flags & LOCATIONS); }
-
-    /** @} */
-
-public:
-
-    /** @name scalar filtering status (experimental; disable at your discretion) */
-    /** @{ */
-
-    /** enable/disable scalar filtering while parsing */
-    ParserOptions& scalar_filtering(bool enabled) noexcept
-    {
-        return set_flags_(enabled, SCALAR_FILTERING);
-    }
-    /** query scalar filtering status */
-    C4_ALWAYS_INLINE bool scalar_filtering() const noexcept { return (flags & SCALAR_FILTERING); }
-
-    /** @} */
 };
